Update checkDuplicateCommands.ts
Array.prototype.join is now being used in the "warnDuplicates" function to avoid nested loops. There is a new function "commandType" to help identify the commands. "checkDuplicateCommands" is now more simplified by creating an obj and adding the CommandFileObjects to an array based on a common command name. Then a loop to push any keys with more than one CommandFileObject. Also if there is only one CommandFileObject in a key it will push that CommandFileObject and not an array of just that one to avoid unnecessary arrays nesting. Because of this change duplicates have been renamed to duplicatePairsArray and nonduplicates have been renamed to nonduplicateArray.